+++ ft_printf.h
@@ -22,7 +22,7 @@ int		ft_putchr(char c);
 int		ft_putstr(char *s);
 int		ft_putnbr(int n);
 int		ft_putunbr(unsigned int n);
-int		ft_putnbr_hex(unsigned int num, int uppercase);
+int		ft_putnbr_hex(unsigned long num, int uppercase);
 int		ft_putptr(void *num);
 
 #endif
